pyproject.toml
[tool.poetry]
name = "domtory"
version = "0.1.0"
description = ""
authors = ["leehjhjhj <tsukiakarii@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
pymysql = "^1.1.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
django-cors-headers = "^4.2.0"
dependency-injector = "^4.41.0"
firebase-admin = "^6.2.0"
bs4 = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
